Feat: refactor SkeletonCard component and update import path
